Reorganize Resource page slightly, unify reservations listings
We don't want to actually show the word "Resource" to the user anywhere
(except maybe just in the address bar), it's a meaningless word and
doesn't make sense when all of the resources are actually rooms, or
something else in particular.

Reservations listings should all be done via the Reservations component,
which will handle fetching and will soon also contain a calendar view.
We may want to remove the root /reservations route anyway, it's probably
not meaningful to combine all resource reservations into a single view.