Handle reservation fetching in List and Add components seperately
It doesn't make sense to only check against the filtered list view when
adding a new event, because it could be outside the filtered range.
Also, we'll add a calendar view soon that will also handle its own event
fetching, so we'll need to do this anyway.

Also, "fix" timezones for the dates returned from WordPress.