Add recurring events
This somewhat naively computes the recurrences on the client in order
to display them to the user for confirmation and to check for conflicts
ahead of time. Recurrences are sent to the server, where it will add
them serially and respond back with which were added and not added.

Needs UX work still.